safari-600.7-branch/LayoutTests/java/lc3/JavaObject/JavaObjectToLong-006-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject to long

FAIL dt.PUB_DOUBLE_REPRESENTATION = -9223372036854777000;dt.setLong( dt ); dt.PUB_LONG should throw an exception. Was 0.
FAIL dt.PUB_DOUBLE_REPRESENTATION = -9223372036854777000;dt.setLong( dt ); dt.getLong() should throw an exception. Was -9223372036854776000.
PASS typeof dt.getLong() is 'number'
PASS successfullyParsed is true

TEST COMPLETE
 
